add new arrays

+++ src/Oksana/NegativeNumbs.java
@@ -5,26 +5,26 @@ import java.util.Random;
 
 public class NegativeNumbs {
     public static void main(String[] args) {
-        double [] array = new double[20];
+        float [] array = new float[20];
         Random r = new Random();
 
         for (int i = 0; i < array.length ; i++)
         {
-            array[i] = r.nextDouble(100+100) - 100;
+            array[i] = r.nextFloat(100+100) - 100;
         }
         int count = 0;
-        for (double value: array)
+        for (float value: array)
         {
-            if ( value < 0 ){
+            if (value < 0 ){
                 count++;
             }
         }
        negative(array, count);
 
     }
-    public static void negative (double [] a, int n){
+    public static void negative (float [] a, int n){
         int negativeCount = 0;
-        double[] b = new double[n];
+        float[] b = new float[n];
 
         for (int i = 0; i < a.length; i++)
         {
@@ -32,9 +32,9 @@ public class NegativeNumbs {
 
                 b[negativeCount]  = a[i];
                 negativeCount++;
-
             }
         }
+
         System.out.println( "Количество отрицательных чисел: " + negativeCount);
         System.out.println(Arrays.toString(a));
         System.out.println(Arrays.toString(b));
